requirements.txt
pymongo>=4.6,<5.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
tenacity>=8.2,<9.0
python-json-logger>=2.0,<3.0
pytest>=7.4,<9.0
mongomock>=4.1,<5.0
ruff>=0.3,<1.0
